refactor(api): move certification  authorization to src
